# server/docs/swagger.yaml
basePath: /
definitions:
...
  types.GroupStatus:
    enum:
    - open
    - closed
    - settled
    type: string
    x-enum-varnames:
    - GroupOpen
    - GroupClosed
    - GroupSettled
  types.GroupSummary:
    properties:
      end_date:
        type: string
      group_id:
        type: string
      member_count:
        type: integer
      members:
        items:
          $ref: '#/definitions/types.MemberSummary'
        type: array
      name:
        type: string
      spend_per_category:
        additionalProperties:
          format: int64
          type: integer
        type: object
      start_date:
        type: string
      status:
        $ref: '#/definitions/types.GroupStatus'
      total_spent:
        type: integer
    type: object
...
  types.MemberBalance:
    properties:
      fair_share:
        type: integer
      net:
        type: integer
      paid:
        type: integer
      user_id:
        type: string
    type: object
  types.MemberSummary:
    properties:
      email:
        type: string
      fair_share:
        type: integer
      net:
        type: integer
      paid:
        type: integer
      user_id:
        type: string
    type: object
...
  types.PaymentStatus:
    enum:
    - pending
    - proof_submitted
    - creditor_confirmed
    - disputed
    - settled
    type: string
    x-enum-varnames:
    - PaymentPending
    - PaymentProofSubmitted
    - PaymentCreditorConfirmed
    - PaymentDisputed
    - PaymentSettled
  types.PaymentView:
    properties:
      amount_baisa:
        type: integer
      created_at:
        type: string
      from:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/types.PaymentStatus'
      to:
        type: string
    type: object
  types.Principal:
    properties:
      email:
        type: string
      id:
        type: string
      is_global_admin:
        type: boolean
      verification_status:
        $ref: '#/definitions/types.VerificationStatus'
    type: object
...
  types.PublicGroupStatus:
    properties:
      member_count:
        type: integer
      name:
        type: string
      status:
        $ref: '#/definitions/types.GroupStatus'
      total_spent:
        type: integer
    type: object
...
  types.SettlementPlanResponse:
    properties:
      group_id:
        type: string
      group_status:
        $ref: '#/definitions/types.GroupStatus'
      note:
        description: Note explains an empty plan (e.g. every net balance was already
          zero).
        type: string
      payments:
        items:
          $ref: '#/definitions/types.PaymentView'
        type: array
      settled_count:
        type: integer
      snapshot:
        allOf:
        - $ref: '#/definitions/types.SettlementSnapshot'
        description: |-
          Snapshot is the state settlement was computed over: total spent, member
          count, and each member's paid / fair share / net.
      total_count:
        type: integer
    type: object
  types.SettlementSnapshot:
    properties:
      member_count:
        type: integer
      members:
        items:
          $ref: '#/definitions/types.MemberBalance'
        type: array
      spend_per_category:
        additionalProperties:
          format: int64
          type: integer
        type: object
      total_spent:
        type: integer
    type: object
...
  types.VerificationStatus:
    enum:
    - registered
    - pending_verification
    - verified
    - rejected
    type: string
    x-enum-varnames:
    - VerificationRegistered
    - VerificationPending
    - VerificationVerified
    - VerificationRejected
  types.apiError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Shared-expense settlement tracker for group trips. Members record what they paid; closing a group computes each member's fair share and the minimum set of payments that squares everyone up. Money moves off-platform — each computed payment is tracked through a two-key confirmation flow (debtor proves, creditor attests, admin finalizes).

    All amounts are integer **baisa** (1.000 OMR = 1000). There are no floats in the money path.

    **Authenticating here:** call `POST /auth/login` (try `admin@expense-splitter.local` / `admin`), copy the `access_token` from the response, click **Authorize** above and paste `Bearer <access_token>`.
  title: Expense Splitter API
  version: "1.0"
paths:
...
  /admin/users/{id}/approve:
    post:
      description: Global admin only. Marks the target account `verified`, unlocking
        group participation.
      parameters:
      - description: user id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the verified account
          schema:
            $ref: '#/definitions/types.Principal'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Approve a user's verification
      tags:
      - admin
  /admin/users/{id}/reject:
    post:
      description: Global admin only. Marks the target account `rejected` — it stays
        blocked from joining groups or recording expenses.
      parameters:
      - description: user id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the rejected account
          schema:
            $ref: '#/definitions/types.Principal'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: global admin only
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Reject a user's verification
      tags:
      - admin
...
  /groups/{id}/settlement:
    get:
      description: |-
        Members and admins, **closed groups only**. The full `{from, to, amount, status}` list produced at close, plus an "N of M settled" progress count. All payments start `pending`.

        The plan is the minimum-ish set of transfers from a greedy largest-debtor/largest-creditor match — at most `n-1` transfers, and it always reconciles to zero.
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the payment plan + progress
          schema:
            $ref: '#/definitions/types.SettlementPlanResponse'
        "400":
          description: invalid group id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not a member of this group
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: group not found
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: group is still open — close it first
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Get the computed payment plan
      tags:
      - settlement
  /groups/{id}/summary:
    get:
      description: 'Members and admins. Trip name and dates, member count, total spent,
        spend per category, and per member: total **paid**, their **fair share**,
        and their **net balance** (paid - fair share). Works on open groups too —
        it is the live picture of who is up and who is down.'
      parameters:
      - description: group id
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: totals, per-category spend, per-member balances
          schema:
            $ref: '#/definitions/types.GroupSummary'
        "400":
          description: invalid group id
          schema:
            $ref: '#/definitions/types.apiError'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: not a member of this group
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: group not found
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Get a group's financial summary
      tags:
      - settlement
...
  /public/groups/{token}:
    get:
      description: '**The only unauthenticated group view** — no bearer token required.
        Given a group''s shareable token, returns a deliberately minimal status (no
        amounts, no member financials) so a link can be shared safely.'
      parameters:
      - description: group share token
        format: uuid
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: minimal public status
          schema:
            $ref: '#/definitions/types.PublicGroupStatus'
        "400":
          description: invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "404":
          description: no group for this token
          schema:
            $ref: '#/definitions/types.apiError'
      summary: Check a group's status by share token
      tags:
      - public
...
  /verification:
    post:
      description: Moves your account `registered` -> `pending_verification`, putting
        it in the global admin's queue. You must reach `verified` before you can join
        a group, record expenses, or be settled against.
      produces:
      - application/json
      responses:
        "200":
          description: account now pending_verification
          schema:
            $ref: '#/definitions/types.Principal'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/types.apiError'
        "403":
          description: caller is not provisioned
          schema:
            $ref: '#/definitions/types.apiError'
        "409":
          description: account is not in a submittable state
          schema:
            $ref: '#/definitions/types.apiError'
      security:
      - BearerAuth: []
      summary: Submit your account for verification
      tags:
      - account
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Bearer token from POST /auth/login. Format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...